back/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
Django==4.2.3
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
docker==7.1.0
faiss-cpu==1.8.0.post1
filelock==3.15.4
FLAML==2.2.0
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
MarkupSafe==2.1.5
marshmallow==3.22.0
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.26.4
openai==1.41.1
orjson==3.10.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pyautogen==0.2.34
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==73.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
sympy==1.13.2
tbb==2021.13.1
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.5
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
